Wallets can now be successfully saved/read from localstorage
