+++ .eslintignore
@@ -8,3 +8,4 @@ babel.config.js
 jest.config.js
 /test
 /types
+quasar.conf.js
