+++ .vscode/settings.json
@@ -13,6 +13,7 @@
   "json.format.enable": true,
   "javascript.format.enable": true,
   "typescript.format.enable": true,
+  "typescript.validate.enable": false,
   "eslint.format.enable": true,
   "vetur.format.defaultFormatter.html": "prettier",
   "vetur.format.defaultFormatter.js": "prettier",
