+++ src/boot/components.js
@@ -1,7 +1,6 @@
 // Configure globally available components
 import BaseButton from 'components/BaseButton';
 import BaseInput from 'components/BaseInput';
-import bitcoreKaspaSetup from 'components/bitcoreKaspaSetup';
 
 // more info on params: https://quasar.dev/quasar-cli/cli-documentation/boot-files#Anatomy-of-a-boot-file
 export default ({ Vue }) => {
@@ -9,5 +8,4 @@ export default ({ Vue }) => {
   /* eslint-disable */
   Vue.component('base-button', BaseButton);
   Vue.component('base-input', BaseInput);
-  bitcoreKaspaSetup();
 };
