+++ src/store/index.ts
@@ -24,7 +24,7 @@ export default store(({ Vue }) => {
 
     // enable strict mode (adds overhead!)
     // for dev mode only
-    strict: !!process.env.DEV,
+    // strict: !!process.env.DEV,
   });
 
   return Store;
