+++ src/wallet/bitcoreKaspaSetup.ts
